% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarad.R
\name{CLARA}
\alias{CLARA}
\title{CLARA clustering}
\usage{
CLARA(x, k, nSamples = 5, sampleFrac = 0.1, swap = FALSE,
  pamonce = 0)
}
\arguments{
\item{x}{(numeric matrix or dist) data}

\item{k}{(positive integer) Number of clusters}

\item{nSamples}{(positive integer, default: 5) Number of random samples}

\item{sampleFrac}{(positive fraction, default: 0.1) Fraction of observations
in a sample}

\item{swap}{(flag, default: FALSE) Whether PAM should involve swap phase}

\item{pamonce}{(One among 0, 1, 2, default: 0) See pamonce argument in
\code{\link[cluster]{pam}}}
}
\value{
A list with three compoments:

  \itemize{

  \item clustering: An integer vector indicating the cluster number with
  length equal to number of observations

  \item medoidsIndex: An integer vector of indices of medoids

  \item cost: average dissimilarity/distance between observations(entire
  dataset) from the nearest medoid

  }
}
\description{
Implements CLARA clustering algorithm using
  \code{\link[cluster]{pam}}
}
\details{
CLARA implementation:

  \itemize{

  \item PAM clustering is computed on multiple random samples of
  observations.

  \item For a given clustering/medoids, cost is defined as the average
  dissimilarity/distance between observations(entire dataset) from the
  nearest medoid.

  \item A clustering/medoids corresponding to the clustering with minimum
  cost is chosen.

  }

  The PAM fitting on multiple subsets is parallelized with \pkg{future}.
}
\examples{
set.seed(1)
clara(dist(mtcars), k = 4, sampleFrac = 0.4, nSamples = 10)
set.seed(2)
clara(stats::dist(mtcars, method = "maximum"), k = 4, sampleFrac = 0.4, nSamples = 10)
}
\author{
Srikanth Komala Sheshachala (sri.teach@gmail.com)
}
